Move extracting parameters from a type out of an Extension
This makes a general implementation of this feature available for every extension which might want to use it.

This can also later be used in several other places internally, instead of having slightly different implementations everywhere.

Not used in SpringSwaggerExtension for now, since it would lead to breaking changes.
