+++ src/main/java/com/github/kongchen/swagger/docgen/reader/AbstractReader.java
@@ -1,75 +1,34 @@
 package com.github.kongchen.swagger.docgen.reader;
 
-import java.lang.annotation.Annotation;
-import java.lang.reflect.Method;
-import java.lang.reflect.Type;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.Iterator;
-import java.util.LinkedHashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-
-import javax.ws.rs.BeanParam;
-import javax.ws.rs.FormParam;
-import javax.ws.rs.HeaderParam;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.QueryParam;
-
-import org.apache.commons.lang3.reflect.TypeUtils;
-import org.apache.commons.lang3.text.StrBuilder;
-import org.apache.maven.plugin.logging.Log;
-import org.springframework.core.annotation.AnnotationUtils;
-import org.springframework.web.bind.annotation.CookieValue;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RequestPart;
-
+import com.github.kongchen.swagger.docgen.util.TypeExtracter;
+import com.github.kongchen.swagger.docgen.util.TypeWithAnnotations;
 import com.google.common.collect.Lists;
 import com.sun.jersey.api.core.InjectParam;
-
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiImplicitParam;
-import io.swagger.annotations.ApiImplicitParams;
-import io.swagger.annotations.ApiOperation;
-import io.swagger.annotations.ApiParam;
-import io.swagger.annotations.ApiResponse;
-import io.swagger.annotations.ApiResponses;
-import io.swagger.annotations.Authorization;
-import io.swagger.annotations.AuthorizationScope;
-import io.swagger.annotations.Extension;
-import io.swagger.annotations.ExtensionProperty;
-import io.swagger.annotations.ResponseHeader;
+import io.swagger.annotations.*;
 import io.swagger.converter.ModelConverters;
 import io.swagger.jaxrs.ext.SwaggerExtension;
 import io.swagger.jaxrs.ext.SwaggerExtensions;
-import io.swagger.models.Model;
-import io.swagger.models.Operation;
+import io.swagger.models.*;
 import io.swagger.models.Path;
-import io.swagger.models.Response;
-import io.swagger.models.Scheme;
-import io.swagger.models.SecurityRequirement;
-import io.swagger.models.Swagger;
 import io.swagger.models.Tag;
-import io.swagger.models.parameters.BodyParameter;
-import io.swagger.models.parameters.FormParameter;
-import io.swagger.models.parameters.HeaderParameter;
-import io.swagger.models.parameters.Parameter;
-import io.swagger.models.parameters.PathParameter;
-import io.swagger.models.parameters.QueryParameter;
+import io.swagger.models.parameters.*;
 import io.swagger.models.properties.ArrayProperty;
 import io.swagger.models.properties.MapProperty;
 import io.swagger.models.properties.Property;
 import io.swagger.models.properties.RefProperty;
 import io.swagger.util.ParameterProcessor;
 import io.swagger.util.PathUtils;
+import org.apache.commons.lang3.reflect.TypeUtils;
+import org.apache.commons.lang3.text.StrBuilder;
+import org.apache.maven.plugin.logging.Log;
+import org.springframework.core.annotation.AnnotationUtils;
+import org.springframework.web.bind.annotation.*;
+
+import javax.ws.rs.*;
+import java.lang.annotation.Annotation;
+import java.lang.reflect.Method;
+import java.lang.reflect.Type;
+import java.util.*;
 
 /**
  * @author chekong on 15/4/28.
@@ -524,6 +483,35 @@ public abstract class AbstractReader {
   		sb.replaceAll("{{httpMethod}}", httpMethod);
   		
   		return sb.toString();
+    }
+
+    public List<Parameter> extractTypes(Class<?> cls, Set<Type> typesToSkip, List<Annotation> additionalAnnotations) {
+        TypeExtracter extractor = new TypeExtracter();
+        Collection<TypeWithAnnotations> typesWithAnnotations = extractor.extractTypes(cls);
+
+        List<Parameter> output = new ArrayList<Parameter>();
+        for (TypeWithAnnotations typeWithAnnotations : typesWithAnnotations) {
+
+            Type type = typeWithAnnotations.getType();
+            List<Annotation> annotations = new ArrayList<Annotation>(additionalAnnotations);
+            annotations.addAll(typeWithAnnotations.getAnnotations());
+
+            /*
+             * Skip the type of the bean itself when recursing into its members
+             * in order to avoid a cycle (stack overflow), as crazy as that user
+             * code would have to be.
+             *
+             * There are no tests to prove this works because the test bean
+             * classes are shared with SwaggerReaderTest and Swagger's own logic
+             * doesn't prevent this problem.
+             */
+            Set<Type> recurseTypesToSkip = new HashSet<Type>(typesToSkip);
+            recurseTypesToSkip.add(cls);
+
+            output.addAll(this.getParameters(type, annotations, recurseTypesToSkip));
+        }
+
+        return output;
     }
 
 	public String getOperationIdFormat() {
